# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Pivot API
  version: 0.1.0
  description: Private invitation-only football dashboard API.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /setup/status:
    get:
      summary: Read whether the first-install assistant is required
      responses: {'200': {description: Installation and setup-token status}}
  /setup:
    post:
      summary: Create the first administrator on an empty installation
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FirstInstall'}
      responses:
        '201': {description: First administrator created}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /auth/register:
    post:
      summary: Register using an invitation code
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Registration'}
      responses:
        '201': {description: Registered}
        '422': {$ref: '#/components/responses/Problem'}
  /auth/login:
    post:
      summary: Start an opaque cookie session
      responses:
        '200': {description: Signed in}
        '401': {$ref: '#/components/responses/Problem'}
  /auth/logout:
    post:
      summary: End the current session
      responses:
        '204': {description: Signed out}
  /auth/me:
    get:
      summary: Read the current user
      security: [{cookieSession: []}]
      responses:
        '200': {description: Current user}
  /profile:
    patch:
      summary: Update nickname and synchronized theme
      security: [{cookieSession: []}]
      responses: {'200': {description: Updated}}
    delete:
      summary: Permanently delete the current account
      security: [{cookieSession: []}]
      responses: {'204': {description: Deleted}}
  /competitions:
    get:
      summary: List locally synchronized competitions
      security: [{cookieSession: []}]
      responses: {'200': {description: Competition catalog and attribution}}
  /clubs:
    get:
      summary: List clubs, optionally filtered by competition code
      parameters:
        - {name: competition, in: query, schema: {type: string}}
      security: [{cookieSession: []}]
      responses: {'200': {description: Club catalog and attribution}}
  /clubs/{id}:
    get:
      summary: Read a club profile
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      security: [{cookieSession: []}]
      responses:
        '200': {description: Club profile and attribution}
        '404': {$ref: '#/components/responses/Problem'}
  /favorites:
    get:
      summary: List the current user's favorite clubs
      security: [{cookieSession: []}]
      responses: {'200': {description: Ordered favorite clubs}}
    put:
      summary: Replace the current user's favorite clubs
      security: [{cookieSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clubIds]
              properties:
                clubIds: {type: array, maxItems: 5, uniqueItems: true, items: {type: string, format: uuid}}
      responses:
        '200': {description: Updated favorite clubs}
        '422': {$ref: '#/components/responses/Problem'}
  /matches:
    get:
      summary: List fixtures and results in a date window
      parameters:
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
        - {name: competition, in: query, schema: {type: string}}
        - {name: club, in: query, schema: {type: string, format: uuid}}
      security: [{cookieSession: []}]
      responses:
        '200': {description: Matches with favorite-club markers and attribution}
        '400': {$ref: '#/components/responses/Problem'}
  /standings:
    get:
      summary: Read the current domestic standing
      parameters:
        - {name: competition, in: query, required: true, schema: {type: string}}
      security: [{cookieSession: []}]
      responses:
        '200': {description: Current total standing and attribution}
        '404': {$ref: '#/components/responses/Problem'}
  /broadcasts:
    get:
      summary: List effective French TV listings in a date window
      parameters:
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
      security: [{cookieSession: []}]
      responses:
        '200': {description: TV listings with linked-match and external markers plus attribution}
        '400': {$ref: '#/components/responses/Problem'}
  /news:
    get:
      summary: List official headlines for favorite clubs or one selected club
      parameters:
        - {name: club, in: query, schema: {type: string, format: uuid}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      security: [{cookieSession: []}]
      responses: {'200': {description: Official-feed metadata retained for 30 days}}
  /admin/invitations:
    get:
      summary: List invitations
      security: [{cookieSession: []}]
      responses: {'200': {description: Invitation list}}
    post:
      summary: Create an invitation and return its code once
      security: [{cookieSession: []}]
      responses: {'201': {description: Created}}
  /admin/collections/football-data:
    get:
      summary: Read connector configuration and latest collection status
      security: [{cookieSession: []}]
      responses: {'200': {description: Connector status}}
    post:
      summary: Run a synchronous catalog collection for the five major leagues
      security: [{cookieSession: []}]
      responses:
        '200': {description: Collection completed}
        '409': {$ref: '#/components/responses/Problem'}
        '502': {$ref: '#/components/responses/Problem'}
  /admin/collections/football-data/sport:
    get:
      summary: Read the latest match and standing collection status
      security: [{cookieSession: []}]
      responses: {'200': {description: Sports collection status}}
    post:
      summary: Synchronize yesterday through the next seven days and current standings
      security: [{cookieSession: []}]
      responses:
        '200': {description: Sports collection completed}
        '409': {$ref: '#/components/responses/Problem'}
        '502': {$ref: '#/components/responses/Problem'}
  /admin/collections/footao:
    get:
      summary: Read the opt-in Footao connector and latest collection status
      security: [{cookieSession: []}]
      responses: {'200': {description: TV collection status}}
    post:
      summary: Synchronize today through the next seven days of French TV listings
      security: [{cookieSession: []}]
      responses:
        '200': {description: TV collection completed}
        '409': {$ref: '#/components/responses/Problem'}
        '502': {$ref: '#/components/responses/Problem'}
  /admin/broadcasts:
    get:
      summary: List TV listings including hidden operator corrections
      security: [{cookieSession: []}]
      responses: {'200': {description: Administrative TV listing view}}
  /admin/broadcasts/{id}/correction:
    put:
      summary: Create or replace an audited operator correction
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      security: [{cookieSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BroadcastCorrection'}
      responses:
        '204': {description: Correction saved and audited}
        '404': {$ref: '#/components/responses/Problem'}
    delete:
      summary: Clear a correction and restore imported values
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      security: [{cookieSession: []}]
      responses:
        '204': {description: Correction cleared and audited}
        '404': {$ref: '#/components/responses/Problem'}
  /admin/broadcasts/audit:
    get:
      summary: List recent TV correction audit entries
      security: [{cookieSession: []}]
      responses: {'200': {description: Most recent 100 audit entries}}
  /admin/news/feeds:
    get:
      summary: List configured official club feeds
      security: [{cookieSession: []}]
      responses: {'200': {description: Feed configuration}}
    post:
      summary: Add or update a public official RSS or Atom feed
      security: [{cookieSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NewsFeedInput'}
      responses:
        '201': {description: Feed saved}
        '422': {$ref: '#/components/responses/Problem'}
  /admin/news/feeds/{id}:
    delete:
      summary: Delete a configured feed and its imported headlines
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      security: [{cookieSession: []}]
      responses:
        '204': {description: Feed deleted}
        '404': {$ref: '#/components/responses/Problem'}
  /admin/collections/news:
    get:
      summary: Read official-feed collection configuration and status
      security: [{cookieSession: []}]
      responses: {'200': {description: News collection status}}
    post:
      summary: Collect configured feeds and enforce 30-day retention
      security: [{cookieSession: []}]
      responses:
        '200': {description: News collection completed}
        '502': {$ref: '#/components/responses/Problem'}
components:
  securitySchemes:
    cookieSession: {type: apiKey, in: cookie, name: pivot_session}
  schemas:
    Registration:
      type: object
      required: [email, password, nickname, invitationCode]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 12}
        nickname: {type: string, minLength: 2, maxLength: 40}
        invitationCode: {type: string}
    FirstInstall:
      type: object
      required: [token, email, password, nickname]
      properties:
        token: {type: string}
        email: {type: string, format: email}
        password: {type: string, minLength: 12, maxLength: 128}
        nickname: {type: string, minLength: 2, maxLength: 40}
    NewsFeedInput:
      type: object
      required: [clubId, url, sourceName, enabled]
      properties:
        clubId: {type: string, format: uuid}
        url: {type: string, format: uri, maxLength: 2000}
        sourceName: {type: string, maxLength: 120}
        enabled: {type: boolean}
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
    BroadcastCorrection:
      type: object
      required: [startsAt, label, kind, channels]
      properties:
        startsAt: {type: string, format: date-time}
        homeName: {type: string, maxLength: 140}
        awayName: {type: string, maxLength: 140}
        label: {type: string, maxLength: 280}
        competitionName: {type: string, maxLength: 140}
        kind: {type: string, enum: [live, delayed, replay]}
        channels: {type: array, maxItems: 12, items: {type: string, maxLength: 100}}
        hidden: {type: boolean}
        note: {type: string, maxLength: 300}
  responses:
    Problem:
      description: RFC 7807 problem
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}